Rename forms/element.ts to forms/pattern.ts
